added drawer, splitted pages in different files and initiated landing pages
